archive/LoadPage/requirements.txt
beautifulsoup4==4.8.1
google-cloud-datastore==1.10.0
requests==2.22.0
